% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_getdata.R
\name{wd_getdata_checks}
\alias{wd_getdata_checks}
\title{Parameter checks}
\usage{
wd_getdata_checks(
  start_dt,
  end_dt,
  tz,
  src,
  units,
  var,
  cache_dir,
  session,
  spinner
)
}
\arguments{
\item{start_dt}{Start date-time}

\item{end_dt}{End date-time}

\item{tz}{Time Zone for the results}

\item{src}{Weather data source / provider}

\item{units}{Units desired (imperial or metric)}

\item{var}{Standardized weather variables}

\item{cache_dir}{Directory for caching}

\item{session}{Shiny session (for showing a spinner)}

\item{spinner}{Show a spinner when fetching data,logical}
}
\value{
TRUE if all checks are passed
}
\description{
Parameter checks
}
\details{
This internal function checks the parameters passed to wd_getdata() and wd_getdata_xxx()
}
