% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_getdata_syn.R
\name{wd_getdata_syn}
\alias{wd_getdata_syn}
\title{Get station data from Synoptic}
\usage{
wd_getdata_syn(
  stid,
  start_dt,
  end_dt,
  var,
  key,
  per = NULL,
  units = NULL,
  tz = Sys.timezone(),
  cache_dir = NULL,
  session = NULL,
  spinner = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{stid}{Station name(s)}

\item{start_dt}{Start date-time}

\item{end_dt}{End date-time}

\item{var}{Standardized weather variables}

\item{key}{API key}

\item{per}{Period / interval (minutes)}

\item{units}{Units desired (imperial or metric)}

\item{tz}{Time Zone for the results}

\item{cache_dir}{Directory for caching}

\item{session}{Shiny session (for showing a spinner)}

\item{spinner}{Show a shinybusy spinner when fetching data, logical}

\item{quiet}{Suppress messages}
}
\value{
A weather data tibble (long format)
}
\description{
Get station data from Synoptic
}
\details{
This will query station data from the SynopticData API.
For documentation, see https://docs.synopticdata.com/services/

To use the Synoptic API, you must create a Public Token. For details log-into your Synoptic account https://customer.synopticdata.com/customer/ and go to 'Create API keys and tokens'.

To cross-check data availability, visit: https://availability.synopticdata.com/stations

The Synoptic time series endpoint (https://docs.synopticdata.com/services/time-series) does not support temporal resampling,
so the \code{per} argument is ignored.

If you pass a value for \code{cache_dir}, downloaded data will be saved in that location. The function however
does not clear the \code{cache_dir} upon closing, so it is recommended you use a temporary directory.
}
