% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_getdata_wwg.R
\name{wd_getdata_wwg}
\alias{wd_getdata_wwg}
\title{Get station data from the Western Weather Group API}
\usage{
wd_getdata_wwg(
  stid,
  start_dt,
  end_dt,
  var,
  key,
  per = NULL,
  units = NULL,
  tz = Sys.timezone(),
  cache_dir = NULL,
  session = NULL,
  spinner = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{stid}{Station name(s)}

\item{start_dt}{Start date-time}

\item{end_dt}{End date-time}

\item{var}{Standardized weather variables}

\item{key}{API key}

\item{per}{Period / interval (minutes)}

\item{units}{Units desired (imperial or metric)}

\item{tz}{Time Zone for the results}

\item{cache_dir}{Directory for caching}

\item{session}{Shiny session (for showing a spinner)}

\item{spinner}{Show a spinner when fetching data, logical}

\item{quiet}{Suppress messages}
}
\value{
A weather data tibble (long format)
}
\description{
Get station data from the Western Weather Group API
}
\details{
This will query station data from the Weather Weather Group API.
For the full documentation, see https://api.westernwx.com/docs/

To use the WWG API, you must have an API key. If your account has API access, you can manage your API Keys from https://app.westernwx.com/apikeys. When you generate a new key, it will give you an Id, Secret, and a Key.

If you pass a value for \code{cache_dir}, downloaded data will be saved in that location. The function however
does not clear the \code{cache_dir} upon closing, so it is recommended you use a temporary directory.
}
