% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_getdata_syn.R
\name{wd_getnetworks_syn}
\alias{wd_getnetworks_syn}
\title{Get Synoptic networks}
\usage{
wd_getnetworks_syn(key, cache = TRUE, quiet = FALSE)
}
\arguments{
\item{key}{API key}

\item{cache}{Cache results}

\item{quiet}{Suppress messages}
}
\description{
Get Synoptic networks
}
