% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_long2wide.R
\name{wd_long2wide}
\alias{wd_long2wide}
\title{Transform a tibble of weather data from long to wide}
\usage{
wd_long2wide(x, use_units = TRUE)
}
\arguments{
\item{x}{A Weather Data tibble}

\item{use_units}{Return columns as units objects}
}
\value{
A weather data tibble (wide format)
}
\description{
Transform a tibble of weather data from long to wide
}
\details{
If \code{units = TRUE}, the columns in the tibble will be of class units from the units package.
}
