% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_vars.R
\name{wd_vars}
\alias{wd_vars}
\title{View standard weather variables}
\usage{
wd_vars()
}
\value{
A tibble of standard weather variables you can use in API calls. Not every API may support every variable.
}
\description{
View standard weather variables
}
\examples{
wd_vars()
}
