% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_getdata.R
\name{wd_getdata}
\alias{wd_getdata}
\title{Download weather station data}
\usage{
wd_getdata(
  src = c("syn", "wwg")[1],
  stid,
  start_dt,
  end_dt,
  var,
  key,
  per = NULL,
  units = NULL,
  tz = Sys.timezone(),
  cache_dir = NULL,
  session = NULL,
  spinner = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{src}{Weather data source / provider}

\item{stid}{Station name(s)}

\item{start_dt}{Start date-time}

\item{end_dt}{End date-time}

\item{var}{Standardized weather variables}

\item{key}{API key}

\item{per}{Period / interval (minutes)}

\item{units}{Units desired (imperial or metric)}

\item{tz}{Time Zone for the results}

\item{cache_dir}{A local directory for caching the results (optional)}

\item{session}{Shiny session (for showing a spinner)}

\item{spinner}{Show a spinner when fetching data,logical}

\item{quiet}{Suppress messages}
}
\value{
A weather data tibble (long format)
}
\description{
Download weather station data
}
\details{
This will query station data from supported APIs.

If you pass a value for \code{cache_dir}, downloaded data will be saved in that location. The function however
does not clear the \code{cache_dir} upon closing, so it is recommended you use a temporary directory.
}
\seealso{
\code{\link{wd_getdata_syn}}, \code{\link{wd_getdata_wwg}}
}
