% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_getdata.R
\name{wd_getdata}
\alias{wd_getdata}
\title{Download weather station data}
\usage{
wd_getdata(
  src = c("syn", "wwg")[1],
  stid,
  start_dt,
  end_dt,
  var,
  key,
  per = NULL,
  units = NULL,
  tz = Sys.timezone()
)
}
\arguments{
\item{src}{Weather data source / provider}

\item{stid}{Station name(s)}

\item{start_dt}{Start date-time}

\item{end_dt}{End date-time}

\item{var}{Standardized weather variables}

\item{key}{API key}

\item{per}{Period / interval (minutes)}

\item{tz}{Time Zone for the results}
}
\value{
A weather data tibble (long format)
}
\description{
Download weather station data
}
\details{
This will query station data from different APIs.
}
\seealso{
\code{\link{wd_getdata_syn}}, \code{\link{wd_getdata_wwg}}
#'
}
