% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_getdata_syn.R
\name{wd_getdata_syn}
\alias{wd_getdata_syn}
\title{Get station data from Synoptic}
\usage{
wd_getdata_syn(
  stid,
  start_dt,
  end_dt,
  var,
  key,
  per = NULL,
  units = NULL,
  tz = Sys.timezone()
)
}
\arguments{
\item{stid}{Station name(s)}

\item{start_dt}{Start date-time}

\item{end_dt}{End date-time}

\item{var}{Standardized weather variables}

\item{key}{API key}

\item{per}{Period / interval (minutes)}

\item{tz}{Time Zone for the results}
}
\value{
A weather data tibble (long format)
}
\description{
Get station data from Synoptic
}
\details{
This will query station data from the SynopticData API.
For documentation, see https://docs.synopticdata.com/services/

To use the Synoptic API, you must create a Public Token. For details log-into your Synoptic account https://customer.synopticdata.com/customer/ and go to 'Create API keys and tokens'.

To cross-check data availability, visit: https://availability.synopticdata.com/stations

The Synoptic time series endpoint (https://docs.synopticdata.com/services/time-series) does not support temporal resampling,
so the \code{per} argument is ignored.
}
